qa/valgrind-suppress: external libbpf.so does not fully initialize attr
The qa/1900 test makes use of libbpf.  When libbpf probes the
linux kernel to check which features are available one of the
probes does not fully initialize all the members of the attr struct.
Adding valgrind suppression entries to qa/1900 to catch
those specific cases.  An upstream issue has been filed on
libbpf to get this addressed:

   https://github.com/libbpf/libbpf/issues/845